pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcap2llm"
version = "0.1.0"
description = "Convert PCAP files into LLM-friendly telecom troubleshooting artifacts."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0",
  "typer>=0.12,<1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5"
]
encrypt = [
  "cryptography>=43.0"
]

[project.scripts]
pcap2llm = "pcap2llm.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pcap2llm = ["profiles/*.yaml"]
